separate out connect button
